backend/requirements.txt
Django~=5.1.0
djangorestframework==3.17.1
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
